Deal with more than one kernel source RPM installed

+++ klibc/lrand48.c
@@ -9,6 +9,6 @@ unsigned short __rand48_seed[3]; /* Common with mrand48.c, srand48.c */
 
 long lrand48(void)
 {
-  return (long)((uint32_t)jrand48(__rand48_seed) >> 1);
+  return (uint32_t)jrand48(__rand48_seed) >> 1;
 }
